Asp.net/feature 01092023 phanchanhvn

+++ Source/GraduateThesisManagement/GraduateThesis.Repository.DAL/ThesisSupervisor.cs
@@ -19,7 +19,7 @@ public partial class ThesisSupervisor
 
     public bool IsCompleted { get; set; }
 
-    public int? Point { get; set; }
+    public decimal? Point { get; set; }
 
     public string Notes { get; set; }
